requirements.txt
tensorflow==2.15.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-binance==1.0.19
